requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest==3.3.1
WebTest==2.0.29
factory-boy==2.9.2

# Management script
Flask-Script==2.0.6

# CORS for api
flask_cors==3.0.3

# Github webhook
Flask-Hookserver==1.1.0

Flask-Admin==1.5.0

# Lint and code style
flake8==3.5.0
flake8-blind-except==0.1.1
flake8-debugger==3.0.0
# flake8-docstrings==1.1.0
flake8-isort==2.2.2
flake8-quotes==0.12.1
isort==4.2.15
pep8-naming==0.4.1
